chatBotOpenAI/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.5.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
openai==1.48.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
